requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.6.1

# LangGraph (Orchestration)
langgraph==0.2.16
langchain-core==0.2.38

# Kafka
kafka-python==2.0.2

# HTTP Client
requests==2.32.3
httpx==0.28.1

# LLM Providers
openai==1.58.1
anthropic==0.42.0
groq==0.13.0

# Git Operations
gitpython==3.1.44

# Environment & Config
python-dotenv==1.0.1

# Logging & Monitoring
prometheus-client==0.21.0
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Utilities
python-multipart==0.0.20
tenacity==8.5.0